Remove implicit finishing of the archive if Builder is dropped.
For a few reasons:

- This allows removing the `'static` requirement on the writer.

- There was no way to abort writing the archive without writing the
  termination sections. If an error happens while building the archive,
  it's probably best to *not* write the termination sections, so that
  the incomplete archive doesn't look complete.

- Writing the termination sections in a separate task made it
  indeterminate when exactly it happened. It could happen long after
  the Builder was dropped.

Writing the termination sections on drop made sense in the original
sync 'tar' crate, where it was more well-defined, but in async world
it's better to require the programmer to be explicit about finishing
it.

This is an incompatible change, if your application doesn't call
`finish` or `into_inner` explicitly.